setup running fastapi dev server locally
