+++ pyproject.toml
@@ -10,6 +10,7 @@ dependencies = [
   "anyio>=4.12.1",
   "async>=0.6.2",
   "bs4>=0.0.2",
+  "fastapi[standard]>=0.136.1",
   "pandas>=3.0.1",
   "playwright==1.50.0",
   "psycopg[binary]>=3.3.3",
