+++ src/static/front.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>My First Webpage</title>
+</head>
+<body>
+    <h1>Welcome to My Website dear {{ name }}</h1>
+    <p>This is a simple paragraph of text on a basic HTML page.</p>
+    <a href="https://www.google.com">Visit Google</a>
+</body>
+</html>
